split helper_functions into 3 files
