+++ frontend/src/helpers/eventHandlers.js
@@ -1,177 +1,6 @@
-import Wkt from "wicket";
 import axios from "axios";
+import { getCurrentAdminTypeURL } from "./reducerHelpers.js";
 
-export const fetchCities = async (setCityURLs) => {
-  
-  const response = await axios.get(`http://localhost:3000/api/0`);
-  // console.log("City Response", response.data)
-  response.data.cityNames.forEach((URL, index) => {
-    const [, cityName] = URL.split("#");
-    
-    setCityURLs((prevCityURLs) => ({
-      ...prevCityURLs,
-      [cityName]: URL,
-    }));
-    
-  });
-};
-
-export const fetchAdministration = async (
-  city,
-  cityURLs,
-  dispatchAdminAreaTypes
-) => {
-  if (city) {
-    try {
-      const response = await axios.post("http://localhost:3000/api/2", {
-        cityName: cityURLs[city],
-      });
-      
-
-
-      // setAdminURLs({ currCity: city }); 
-      dispatchAdminAreaTypes({
-        type: "SET_CURRENT_CITY",
-        payload: city
-      });
-      const adminAreaTypeURLs = {};
-
-      // console.log("admin areas", response.data.adminAreaTypeNames);
-      response.data.adminAreaTypeNames.forEach((URL, index) => {
-        const [, adminName] = URL.split("#");
-
-        adminAreaTypeURLs[adminName] = { URL: URL, selected: false };
-      });
-
-      dispatchAdminAreaTypes({
-        type: "SET_URLS",
-        payload: adminAreaTypeURLs
-      });
-
-    } catch (error) {
-      console.error("POST Error:", error);
-    }
-  } 
-};
-
-export const fetchIndicators = async (
-  city,
-  cityURLs,
-  setIndicatorURLs
-) => {
-  if (city) {
-    try {
-      const response = await axios.post("http://localhost:3000/api/1", {
-        cityName: cityURLs[city],
-      });
-      // console.log("indicators", response.data.indicatorNames);
-      
-      response.data.indicatorNames.forEach((URL, index) => {
-        const [, indName] = URL.split("#");
-
-        setIndicatorURLs((prevIndicatorURLs) => ({
-          ...prevIndicatorURLs,
-          [indName]: URL,
-        }));
-      });
-    } catch (error) {
-      console.error("POST Error:", error);
-    }
-  } else {
-    setIndicatorURLs({});
-  }
-};
-
-export const fetchArea = async (
-  admin,
-  cityURLs,
-  adminAreaTypesState,
-  setAreaURLs,
-  setCurrentAreaNames
-) => {
-  setAreaURLs({});
-  setCurrentAreaNames({});
-  if (admin) {
-    try {
-      const areaTypeURL = adminAreaTypesState[admin].URL;
-      const cityName = cityURLs[adminAreaTypesState["currCity"]];
-
-      const response = await axios.post("http://localhost:3000/api/3", {
-        cityName: cityName,
-        adminType: areaTypeURL,
-      });
-      console.log("admin instances", response.data["adminAreaInstanceNames"]);
-
-      response.data["adminAreaInstanceNames"].forEach((Instance, index) => {
-        setAreaURLs((prevAreaURLs) => ({
-          ...prevAreaURLs,
-          [Instance["areaName"]]: Instance["adminAreaInstance"],
-        }));
-
-        setCurrentAreaNames((prevCurrent) => ({
-          ...prevCurrent,
-          [Instance["adminAreaInstance"]]: Instance["areaName"],
-        }));
-      });
-    } catch (error) {
-      console.error("POST Error:", error);
-    }
-  } else {
-    setAreaURLs({});
-    setCurrentAreaNames({});
-  }
-};
-
-export const fetchLocations = async (
-  admin,
-  cityURLs,
-  adminAreaTypesState,
-  locationURLs,
-  setLocationURLs
-) => {
-  setLocationURLs({});
-  if (admin) {
-    try {
-      const response = await axios.post("http://localhost:3000/api/6", {
-        cityName: cityURLs[adminAreaTypesState["currCity"]],
-        adminType: adminAreaTypesState[admin].URL,
-      });
-      console.log("list of all admin instances", response.data["adminAreaInstanceNames"]);
-
-      const updatedLocationURLs = { ...locationURLs };
-      response.data["adminAreaInstanceNames"].forEach((Instance, index) => {
-        var wkt = new Wkt.Wkt();
-        wkt.read(Instance["areaLocation"]);
-
-        var flipped = wkt.toJson();
-
-        // The coordinates are FLIPPED in the database (Lon/Lat instead of Lat/Lon).
-        // The code requires Lat/Lon, so flip it back.
-
-        if (flipped.type === "Polygon") {
-          flipped.coordinates = flipped.coordinates.map((innerArray) =>
-            innerArray.map((coords) => [coords[1], coords[0]])
-          );
-        } else {
-          // flipped is a MULTIpolygon
-          flipped.coordinates = flipped.coordinates.map((firstInnerArray) => 
-            firstInnerArray.map((secondInnerArray) => 
-              secondInnerArray.map((coords) => [coords[1], coords[0]])
-            )
-          );
-        }
-        
-        updatedLocationURLs[Instance["adminAreaInstance"]] = flipped;
-      });
-      setLocationURLs(updatedLocationURLs);
-      console.log("locations", updatedLocationURLs);
-    } catch (error) {
-      console.error("POST Error:", error);
-    }
-  } else {
-    setLocationURLs({});
-  }
-};
 export const handleSum = (indicator, chartData) => {
   let data = JSON.parse(JSON.stringify(chartData[indicator]));
   for (const yearData of data) {
@@ -220,6 +49,7 @@ export const handleAggregation = (indicator, chartData) => {
   const aggregatedArray = Object.values(aggregatedData);
   return aggregatedArray;
 };
+
 export const handleChangeAreas = (
   event,
   setCurrentAdminInstances,
@@ -275,7 +105,7 @@ export const handleUpdateYear = (id, startOrEnd, event, years, setYears) => {
   }
   var sliced_years = years.slice(id + 1);
   if (sliced_years.length !== 0) {
-    for(var y in sliced_years){
+    for (var y in sliced_years) {
       temp.push(sliced_years[y]);
     }
     // temp.push(years.slice(id + 1));
@@ -288,16 +118,6 @@ export const handleUpdateIndicators = (id, value, setSelectedIndicators) => {
     ...prevState,
     [id]: value,
   }));
-};
-
-//TODO: Move this function to a more appropriate location
-export const getCurrentAdminTypeURL = (adminAreaTypesState) => {
-  for (const key in adminAreaTypesState) {
-    if (adminAreaTypesState[key].selected) {
-      return adminAreaTypesState[key].URL;
-    }
-  }
-  return null;
 };
 
 
@@ -320,7 +140,6 @@ export const handleGenerateVisualization = async (
 
   const checkIfInputsFilled = () => {
     return (
-
       typeof adminAreaTypesState["currCity"] !== "undefined" &&
       typeof currentAdminType === "string" &&
       currentAdminType !== "" &&
@@ -352,43 +171,42 @@ export const handleGenerateVisualization = async (
         startTime: years[parseInt(index)].value1,
         endTime: years[parseInt(index)].value2,
       });
-  
+
       const unitType = await axios.post("http://localhost:3000/api/5", {
         subject: indicatorURLs[selectedIndicators[index]],
-        predicate: "http://ontology.eil.utoronto.ca/ISO21972/iso21972#hasUnit"
+        predicate: "http://ontology.eil.utoronto.ca/ISO21972/iso21972#hasUnit",
       });
-  
+
       if (unitType.data["propertyValue"].length === 0) {
-        
         unitType.data["propertyValue"].push("NONE");
       }
-  
+
       console.log(
         "final data",
         index,
         unitType.data["propertyValue"],
-        response.data["indicatorDataValues"],
+        response.data["indicatorDataValues"]
       );
-  
+
       return {
         indicator: indicatorURLs[selectedIndicators[index]],
         data: response.data["indicatorDataValues"],
-        unit: unitType.data["propertyValue"]
+        unit: unitType.data["propertyValue"],
       };
     });
-  
+
     const results = await Promise.all(promises);
-  
+
     const newData = results.reduce((acc, { indicator, data, unit }) => {
       acc[indicator] = { data, unit };
       return acc;
     }, {});
-  
+
     setIndicatorData((prevData) => ({
       ...prevData,
-      ...newData
+      ...newData,
     }));
-  }
+  };
 
   if (checkIfInputsFilled()) {
     if (showVisError) {
@@ -448,8 +266,4 @@ export const handleGenerateVisualization = async (
     setShowingVisualization(false);
   }
   setVisLoading(false);
-};
-
-export const addSavedIndicator = () => {
-
 };
