+++ frontend/src/helpers/reducerHelpers.js
@@ -1,0 +1,9 @@
+// ## AdminAreaTypesReducer helper functions
+export const getCurrentAdminTypeURL = (adminAreaTypesState) => {
+  for (const key in adminAreaTypesState) {
+    if (adminAreaTypesState[key].selected) {
+      return adminAreaTypesState[key].URL;
+    }
+  }
+  return null;
+};
